+++ Calculations/Interface/ICalculation.cs
@@ -12,13 +12,13 @@ namespace TradingApplication___Console.Calculations.Interface
 
         float CalculateRelativeStrength(List<EOD> dailyEODs, int period);
         float CalculateResults(List<EOD> dailyEODs, int period);
-        float CalculateMovingAverage(List<EOD> dailyEODs, int period);
+        decimal CalculateMovingAverage(List<EOD> dailyEODs, int period);
         float CalculateRelativeStrengthVolume(List<EOD> dailyEODs, int period);
 
         Task<float> CalculateVolatilityAsync(List<EOD> dailyEODs, int period);
         Task<float> CalculateRelativeStrengthAsync(List<EOD> dailyEODs, int period);
         Task<float> CalculateResultsAsync(List<EOD> dailyEODs, int period);
-        Task<float> CalculateMovingAverageAsync(List<EOD> dailyEODs, int period);
+        Task<decimal> CalculateMovingAverageAsync(List<EOD> dailyEODs, int period);
 
 
     }
